data-managament/pyqt/requirements.txt
PyQt6>=6.9.0
psycopg2-binary>=2.9.0
